refactor(store): own persisted signal vocabulary
