+++ crates/rimz/src/store/follow.rs
@@ -7,10 +7,9 @@ use std::io;
 use crate::agents::AgentState;
 use crate::agents::lifecycle::{LifecycleEvent, LifecycleSignal, LifecycleState, step};
 use crate::disk::paths::StatePaths;
-use crate::harness::schedule::signal::{SignalName, SignalSource};
 use crate::ids::{AgentKind, AgentSessionId};
 use crate::ids::{EventId, WorkspaceId};
-use crate::store::event::EventKind;
+use crate::store::event::{EventKind, SignalName, SignalSource};
 use crate::store::{event_log, snapshot};
 use jiff::Timestamp;
 use serde::Serialize;
