+++ refactor-target.toml
@@ -244,7 +244,7 @@ upward-dependencies = [
     "harness::target",
     "remote::link",
 ]
-surface-budget = 316
+surface-budget = 323
 
 [[module]]
 path = "crates/rimz/src/theme"
@@ -684,7 +684,7 @@ reason = "Reached by inference from the binary; narrowing is a private-interface
 
 [[verdict]]
 kind = "item"
-key = "harness::schedule::signal::SignalNameErr"
+key = "store::event::SignalNameErr"
 reason = "Reached by inference from the binary; narrowing is a private-interface lint."
 
 [[verdict]]
